Log rate limit information for Octokit
